always check if there are missing icons

+++ scripts/beforePack.js
@@ -9,34 +9,33 @@ const createTrayImage = (name, opt, mult) => new Promise(async res => {
     const regularPath = outputFilePath(`${name}`, mult);
     const invertedPath = outputFilePath(`${name}-inv`, mult);
 
-    console.log(`getting buffer of ${name} @ ${opt.width}`)
+    if(!fs.existsSync(regularPath)) {
+        console.log(`getting buffer of ${name} @ ${mult} (${opt.width})`);
 
-    const buf = fs.readFileSync('res/trayIcons/' + name + '.svg');
-    const png = await convert(buf, opt);
+        const buf = fs.readFileSync('res/trayIcons/' + name + '.svg');
+        const png = await convert(buf, opt);
+        
+        fs.writeFileSync(regularPath, png);
+    };
+
+    if(!fs.existsSync(invertedPath)) {
+        console.log(`getting inverted buffer of ${name} @ ${mult} (${opt.width})`)
     
-    fs.writeFileSync(regularPath, png);
-
-    console.log(`getting inverted buffer of ${name} @ ${opt.width}`)
-
-    const inv = await jimp.read(png);
-    inv.invert();
-    inv.getBuffer(jimp.MIME_PNG, (err, invBuf) => {
-        if(!err) {
-            fs.writeFileSync(invertedPath, invBuf);
-            res();
-        } else throw new Error(`Failed to invert ${name} @ ${opt.width}`, err)
-    });
+        const inv = await jimp.read(png);
+        inv.invert();
+        inv.getBuffer(jimp.MIME_PNG, (err, invBuf) => {
+            if(!err) {
+                fs.writeFileSync(invertedPath, invBuf);
+                res();
+            } else throw new Error(`Failed to invert ${name} @ ${mult} (${opt.width})`, err)
+        });
+    } else res();
 })
 
 module.exports = (context) => new Promise(async res => {
-    let sizes = [ 16 ];
-
     let supportedMultipliers = [ 1, 1.25, 1.33, 1.4, 1.5, 1.8, 2, 2.5, 3, 4, 5 ]
-    //https://github.com/electron/electron/blob/main/docs/api/native-image.md#template-image
-
-    for(mult of supportedMultipliers.slice(1)) {
-        sizes.push( 16 * mult );
-    }
+    let sizes = supportedMultipliers.map(m => 16 * m);
+    // https://www.electronjs.org/docs/latest/api/native-image#high-resolution-image
 
     fs.mkdirSync(`dist/trayIcons`, { recursive: true });
